Update ikfk2Bone.cpp
Changed the way the softness was calculated to work no matter if the stretch is on or off. The stretch had to be altered to work along side and in conjunction with the softness. The max stretch should be changed in the attribute editor from 1 to 99 to 0 to 1 since it is now a percentage. This is easily fixable with editing the attributes or changing this in shifter.